+++ src/cg_cgame.hpp
@@ -16,10 +16,10 @@ struct cg_t {
 };
 
 void  CG_Init();
-cg_t& CG_GetLocalClientGlobals(int localClientNum);
-bool  CG_LocalClientIsActive  (int localClientNum);
-void  CG_ActivateLocalClient  (int localClientNum);
-void  CG_DectivateLocalClient (int localClientNum);
+cg_t& CG_GetLocalClientGlobals(size_t localClientNum);
+bool  CG_LocalClientIsActive  (size_t localClientNum);
+void  CG_ActivateLocalClient  (size_t localClientNum);
+void  CG_DectivateLocalClient (size_t localClientNum);
 
 void  CG_Frame(uint64_t deltaTime);
 void  CG_Shutdown();
